% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/km2gcn.R
\name{applykM2WGCNA}
\alias{applykM2WGCNA}
\title{This is the main function of the package. It takes a hopefully correctly created WGCN network and,
based on gene expression it modifies the module color assignment by using k-means heuristic.
The description of the parameters show how the function works.}
\usage{
applykM2WGCNA(net.label, net.file, expr.data, beta = 6, n.iterations = 20,
  meg = 20, job.path, tom.matrix = NULL, plot.evolution = TRUE,
  plot.go = plot.evolution, debug = NULL, net.type = "signed",
  final.net = paste0(job.path, "/", net.label, ".", beta, ".", net.type,
  ".it.", n.iterations, ".rds"), min.genes.for.grey = 20)
}
\arguments{
\item{net.label}{This parameter is useful for referring to this specific network at any plot that
might be created to monitor the algorithm's behavior}

\item{net.file}{can be either a full path to a RDS object storing the network or the network itself.
We expect a list with objects "moduleColors" and "MEs". The "moduleColors" element si a named vector
whose elements are colors as assigned by WGCNA. The names() function should return the gene names for
each vector's element. The "MEs" is a matrix with a column for each module and a row for each sample.
It contains the module eigengenes. Note that the order of genes in moduleColors is the same as they
appear in columns in expr.data parameter's object.}

\item{expr.data}{is, again, a full path file name or a data frame with genes in columns and samples in
rows, giving the expression data used to construct the WGCNA network. The function expects genes in the
column order to correspond to the order of nameColors of net.file param.}

\item{beta}{The soft threshold parameter as it was used with WGCNA}

\item{n.iterations}{A max number of iterations to run the k-means}

\item{job.path}{Where to store the files generated by the function}

\item{tom.matrix}{If the TOM is precalculated, it can be pased as full path file name or as an object.}

\item{plot.evolution}{If TRUE, it will generate a full set of plots that help making a proper diagnose of
the function.}

\item{plot.go}{If plot.evolution is TRUE and this is also TRUE, it will include in the evolution plot an
evaluation of the GO (GEne Ontology) signals obtained from the new partition, in comparision with the
original partition obtained by the standard WGCNA}

\item{debug}{It can be used to set a small number of genes just to run the fuction tentatively before using
the whole gene set. May be useful to get a flavour on how the method works.}

\item{net.type}{it is used in the same sense as WGCNA}

\item{final.net}{where to store as an RDS object the final network obtained}

\item{min.genes.for.grey}{The number of genes to consider grey genes as part of the optimization}
}
\value{
The network, post-processed with a k-means heuristic
}
\description{
This is the main function of the package. It takes a hopefully correctly created WGCN network and,
based on gene expression it modifies the module color assignment by using k-means heuristic.
The description of the parameters show how the function works.
}

