% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotEvolution.R
\name{plotEvolution}
\alias{plotEvolution}
\title{This function can be used to visualice the results of a k-means heuristic applied
to a WGCNA network. It can be called internally by \code{applykM2WGCNA} or
independently.}
\usage{
plotEvolution(net.label, net, module.colors = unique(net$moduleColors),
  partitions, tom = NULL, beta, expr.data, net.type, plot.evolution.file,
  modules = NULL, go.file)
}
\arguments{
\item{net.label}{A label to use within the plots, to refer to the GCN}

\item{net}{The network obtained by the k-means process (WGCNA object)}

\item{module.colors}{Final partition (the default will be usually fine)}

\item{partitions}{The RDS partitions file as generated by \code{applykM2WGCNA}.}

\item{tom}{The TOM (Topology Overlap Matrix) created by WGCNA, or reproduced
from the combination of expression data + beta}

\item{beta}{The soft thresholding used for the GCN in WGCNA}

\item{expr.data}{The data frame with gene expression (genes in columns, samples in rows)}

\item{net.type}{The net type as used in WGCNA}

\item{plot.evolution.file}{The pdf generated with all the plots}

\item{modules}{Just if we wanted to focus on specific modules, a list of strings}

\item{go.file}{One important evaluation of the optimization is the number of Gene Ontology
significant terms we get from the modules. If go.file is not NULL, a gProfiler based
analysis result will be saved on this file as a rds object. Results will be plotted within
the general plot.}
}
\value{
None
}
\description{
This function can be used to visualice the results of a k-means heuristic applied
to a WGCNA network. It can be called internally by \code{applykM2WGCNA} or
independently.
}

