% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGOenrichment.R
\name{plotGOenrichment}
\alias{plotGOenrichment}
\title{This function can be used to visualice the number of GO (Gene Ontology) signals
gained and lost between the WGCNA standard network and the new network obtained
from the application of k-means search heuristic. It is based on the gProfileR
package}
\usage{
plotGOenrichment(net.label, net, partitions, plot.file = NULL,
  filter = "GO", exclude.iea = T, gprof.method = "gSCS")
}
\arguments{
\item{net.label}{A label to use within the plots, to refer to the GCN}

\item{net}{The network obtained by the k-means process (WGCNA object)}

\item{partitions}{The partition file}

\item{plot.file}{If not NULL, then it should be the pdf file name in which the
GO gain/lost will be presented.}

\item{filter}{The filter used for gProfileR (see details at this package)}

\item{exclude.iea}{If true, do not use Inferred Electronic Annotations in the
enrichment analysis (gProfileR)}

\item{gprof.method}{The default is the one implemented by gProfiler (see package
for details)}
}
\value{
The results from the analysis for further study
}
\description{
This function can be used to visualice the number of GO (Gene Ontology) signals
gained and lost between the WGCNA standard network and the new network obtained
from the application of k-means search heuristic. It is based on the gProfileR
package
}

