% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDistance.R
\name{corDistance}
\alias{corDistance}
\title{Function that generates the correlation for distance, in line
with a signed network}
\usage{
corDistance(a, b, signed = TRUE)
}
\arguments{
\item{a}{numerical vector with gene expression for gene a}

\item{b}{numerical vector with gene expression for gene a}

\item{signed}{is the network signed?}
}
\value{
The signed correlation if signed is TRUE
}
\description{
Function that generates the correlation for distance, in line
with a signed network
}

